feat: add custom navbar/header's for error pages
The header component was moved to render from the individual pages
so that error pages can specify their own custom header instead
of depending on the header rendered in _app.tsx.

The PawnShopHeader component must now be wrapped so it's no longer
exported. Only the provided wrapped headers should be used.