Update Fieldset.module.css
We added 'word-break: break-all' earlier so that long NFT titles didn't break the "Collateral" section layout. Unfortunately this was *too much breaking* and now we have labels (ex. PAYBACK AT MATURITY) that break in the middle of words (viz. PAYBACK AT MAT      URITY). The correct CSS for what we want is `break-word` but that has been deprecated in favor of 'word- break: normal;' combined with 'overflow-wrap: anywhere'. This PR adds that to our fieldset module.